add InGame.java
